fix(pwa): preserve tricount part counts
