+++ .changeset/steady-tricount-parts.md
@@ -1,0 +1,5 @@
+---
+"@trizum/pwa": patch
+---
+
+Preserve Tricount ratio part counts during import so expenses split by parts reopen with the original participant counts in the editor.
